slides/requirements.txt
python-pptx>=0.6.21
pathlib>=1.0.1
pillow>=9.0.0  # for image handling 